#Handlers Refactor next part

+++ src/Application/Features/ExampleLinks/Commands/DeleteExampleLink.cs
@@ -1,8 +1,12 @@
 ﻿using Application.Abstractions;
 using Application.Abstractions.IRepository;
+using Application.Extensions;
 using Domain.Entities.MidjourneyStyles;
 using Domain.ValueObjects;
 using FluentResults;
+using static Application.Errors.ApplicationErrorMessages;
+using static Domain.Errors.DomainErrorMessages;
+using Domain.Errors;
 
 namespace Application.Features.ExampleLinks.Commands;
 
@@ -17,7 +21,25 @@ public static class DeleteExampleLink
 
         public async Task<Result<MidjourneyStyleExampleLink>> Handle(Command command, CancellationToken cancellationToken)
         {
-            await Validate.Link.ShouldExists(command.Link, _exampleLinkRepository);
+            List<DomainError> domainErrors = [];
+
+            domainErrors
+                .CollectErrors<ExampleLink>(command.Link);
+
+            List<ApplicationError> applicationErrors = [];
+
+            applicationErrors
+                .IfLinkNotExists(command.Link, _exampleLinkRepository);
+
+            if (applicationErrors.Count != 0 || domainErrors.Count != 0)
+            {
+                var error = new Error("Validation failed")
+                    .WithMetadata("Application Errors", applicationErrors)
+                    .WithMetadata("Domain Errors", domainErrors);
+
+                return Result.Fail<MidjourneyStyleExampleLink>(error);
+            }
+
 
             return await _exampleLinkRepository.DeleteExampleLinkAsync(command.Link);
         }
