+++ src/Application/Features/Properties/Commands/DeletePropertyInVersion.cs
@@ -1,12 +1,18 @@
 using Application.Abstractions;
 using Application.Abstractions.IRepository;
+using Application.Extensions;
+using Domain.Entities.MidjourneyStyles;
+using Domain.Errors;
+using Domain.ValueObjects;
 using FluentResults;
+using static Application.Errors.ApplicationErrorMessages;
+using static Domain.Errors.DomainErrorMessages;
 
 namespace Application.Features.Properties.Commands;
 
 public static class DeletePropertyInVersion
 {
-    public sealed record Command(string Version, string PropertyName) : ICommand<PropertyDetails>;
+    public sealed record Command(ModelVersion Version, PropertyName PropertyName) : ICommand<PropertyDetails>;
 
     public sealed class Handler(IVersionRepository versionRepository, IPropertiesRepository propertiesRepository)
         : ICommandHandler<Command, PropertyDetails>
@@ -16,8 +22,26 @@ public static class DeletePropertyInVersion
 
         public async Task<Result<PropertyDetails>> Handle(Command command, CancellationToken cancellationToken)
         {
-            await Validate.Version.ShouldExists(command.Version, _versionRepository);
-            await Validate.Property.ShouldExists(command.Version, command.PropertyName, _propertiesRepository);
+            List<DomainError> domainErrors = [];
+
+            domainErrors
+                .CollectErrors<ModelVersion>(command.Version)
+                .CollectErrors<PropertyName>(command.PropertyName);
+
+            List<ApplicationError> applicationErrors = [];
+
+            applicationErrors
+                .IfVersionNotExists(command.Version, _versionRepository)
+                .IfPropertyNotExists(command.Version, command.PropertyName, _propertiesRepository);
+
+            if (applicationErrors.Count != 0 || domainErrors.Count != 0)
+            {
+                var error = new Error("Validation failed")
+                    .WithMetadata("Application Errors", applicationErrors)
+                    .WithMetadata("Domain Errors", domainErrors);
+
+                return Result.Fail<PropertyDetails>(error);
+            }
 
             return await _propertiesRepository.DeleteParameterInVersionAsync(command.Version, command.PropertyName);
         }
